Moved darkmode logic to seperate component
